+++ packages/common/src/index.ts
@@ -67,11 +67,17 @@ import {
     WarehouseCredentials,
 } from './types/projects';
 import { ResultRow } from './types/results';
-import { SchedulerAndTargets, SchedulerWithLogs } from './types/scheduler';
+import {
+    ApiJobScheduledResponse,
+    ApiJobStatusResponse,
+    SchedulerAndTargets,
+    SchedulerWithLogs,
+} from './types/scheduler';
 import { SlackChannel } from './types/slack';
 import { Space } from './types/space';
 import { TableBase } from './types/table';
 import { LightdashUser, UserAllowedOrganization } from './types/user';
+import { ValidationResponse } from './types/validation';
 import { formatItemValue } from './utils/formatting';
 import { getItemId, getItemLabelWithoutTableName } from './utils/item';
 
@@ -479,7 +485,10 @@ type ApiResults =
     | ApiScheduledDownloadCsv
     | PinnedItems
     | ViewStatistics
-    | SchedulerWithLogs;
+    | SchedulerWithLogs
+    | ValidationResponse[]
+    | ApiJobStatusResponse['results']
+    | ApiJobScheduledResponse['results'];
 
 export type ApiResponse = {
     status: 'ok';