+++ packages/frontend/src/components/UserSettings/AccessTokensPanel/index.tsx
@@ -12,37 +12,36 @@ const AccessTokensPanel: FC = () => {
     const [isCreatingToken, setIsCreatingToken] = useState(false);
     const hasAvailableTokens = data && data.length > 0;
 
-    if (!hasAvailableTokens) {
-        return (
-            <EmptyState
-                icon={
-                    <MantineIcon
-                        icon={IconKey}
-                        color="gray.6"
-                        stroke={1}
-                        size="5xl"
-                    />
-                }
-                title="No tokens"
-                description="You haven't generated any tokens yet!, generate your first token"
-            >
-                <Button onClick={() => setIsCreatingToken(true)}>
-                    Generate token
-                </Button>
-            </EmptyState>
-        );
-    }
-
     return (
         <Stack mb="lg">
-            <Group position="apart">
-                <Title order={5}>Personal access tokens</Title>
-                <Button onClick={() => setIsCreatingToken(true)}>
-                    Generate new token
-                </Button>
-            </Group>
-
-            <TokensTable />
+            {hasAvailableTokens ? (
+                <>
+                    <Group position="apart">
+                        <Title order={5}>Personal access tokens</Title>
+                        <Button onClick={() => setIsCreatingToken(true)}>
+                            Generate new token
+                        </Button>
+                    </Group>
+                    <TokensTable />
+                </>
+            ) : (
+                <EmptyState
+                    icon={
+                        <MantineIcon
+                            icon={IconKey}
+                            color="gray.6"
+                            stroke={1}
+                            size="5xl"
+                        />
+                    }
+                    title="No tokens"
+                    description="You haven't generated any tokens yet!, generate your first token"
+                >
+                    <Button onClick={() => setIsCreatingToken(true)}>
+                        Generate token
+                    </Button>
+                </EmptyState>
+            )}
 
             {isCreatingToken && (
                 <CreateTokenModal